clone before set — fixes #479
