+++ test/runtime/samples/set-clones-input/main.html
@@ -1,0 +1,7 @@
+<script>
+	export default {
+		computed: {
+			b: a => a + 1
+		}
+	};
+</script>
